switch from global variables to function parameters in `Yoga`

+++ yoga/YGLayout.h
@@ -22,12 +22,12 @@ struct YGLayout {
   bool doesLegacyStretchFlagAffectsLayout : 1;
   bool hadOverflow : 1;
 
-  uint32_t computedFlexBasisGeneration = 0;
+  uint8_t computedFlexBasisGeneration = 0;
   YGFloatOptional computedFlexBasis = {};
 
   // Instead of recomputing the entire layout every single time, we cache some
   // information to break early when nothing changed
-  uint32_t generationCount = 0;
+  uint8_t generationCount = 0;
   YGDirection lastOwnerDirection = (YGDirection) -1;
 
   uint32_t nextCachedMeasurementsIndex = 0;
